# 422 handle in_keys in booster if possible