--- a/test/rpc-integration/resources/collectiontest/singleton-map.json
+++ b/test/rpc-integration/resources/collectiontest/singleton-map.json
@@ -1,0 +1,167 @@
+{
+  "format": "KORE",
+  "version": 1,
+  "term": {
+    "tag": "App",
+    "name": "Lbl'-LT-'generatedTop'-GT-'",
+    "sorts": [],
+    "args": [
+      {
+        "tag": "App",
+        "name": "Lbl'-LT-'k'-GT-'",
+        "sorts": [],
+        "args": [
+          {
+            "tag": "App",
+            "name": "kseq",
+            "sorts": [],
+            "args": [
+              {
+                "tag": "App",
+                "name": "inj",
+                "sorts": [
+                  {
+                    "tag": "SortApp",
+                    "name": "SortState",
+                    "args": []
+                  },
+                  {
+                    "tag": "SortApp",
+                    "name": "SortKItem",
+                    "args": []
+                  }
+                ],
+                "args": [
+                  {
+                    "tag": "App",
+                    "name": "LblState",
+                    "sorts": [],
+                    "args": [
+                      {
+                        "tag": "App",
+                        "name": "inj",
+                        "sorts": [
+                          {
+                            "tag": "SortApp",
+                            "name": "SortMap",
+                            "args": []
+                          },
+                          {
+                            "tag": "SortApp",
+                            "name": "SortCollection",
+                            "args": []
+                          }
+                        ],
+                        "args": [
+                              {
+                                "tag": "App",
+                                "name": "Lbl'UndsPipe'-'-GT-Unds'",
+                                "sorts": [],
+                                "args": [
+                                  {
+                                    "tag": "App",
+                                    "name": "inj",
+                                    "sorts": [
+                                      {
+                                        "tag": "SortApp",
+                                        "name": "SortInt",
+                                        "args": []
+                                      },
+                                      {
+                                        "tag": "SortApp",
+                                        "name": "SortKItem",
+                                        "args": []
+                                      }
+                                    ],
+                                    "args": [
+                                      {
+                                        "tag": "DV",
+                                        "sort": {
+                                          "tag": "SortApp",
+                                          "name": "SortInt",
+                                          "args": []
+                                        },
+                                        "value": "6"
+                                      }
+                                    ]
+                                  },
+                                  {
+                                    "tag": "App",
+                                    "name": "inj",
+                                    "sorts": [
+                                      {
+                                        "tag": "SortApp",
+                                        "name": "SortBool",
+                                        "args": []
+                                      },
+                                      {
+                                        "tag": "SortApp",
+                                        "name": "SortKItem",
+                                        "args": []
+                                      }
+                                    ],
+                                    "args": [
+                                      {
+                                        "tag": "App",
+                                        "name": "Lblf",
+                                        "sorts": [],
+                                        "args": [
+                                          {
+                                            "tag": "DV",
+                                            "sort": {
+                                              "tag": "SortApp",
+                                              "name": "SortInt",
+                                              "args": []
+                                            },
+                                            "value": "1"
+                                          }
+                                        ]
+                                      }
+                                    ]
+                                  }
+                                ]
+                          }
+                        ]
+                      },
+                      {
+                        "tag": "DV",
+                        "sort": {
+                          "tag": "SortApp",
+                          "name": "SortInt",
+                          "args": []
+                        },
+                        "value": "5"
+                      }
+                    ]
+                  }
+                ]
+              },
+              {
+                "tag": "App",
+                "name": "dotk",
+                "sorts": [],
+                "args": []
+              }
+            ]
+          }
+        ]
+      },
+      {
+        "tag": "App",
+        "name": "Lbl'-LT-'generatedCounter'-GT-'",
+        "sorts": [],
+        "args": [
+          {
+            "tag": "DV",
+            "sort": {
+              "tag": "SortApp",
+              "name": "SortInt",
+              "args": []
+            },
+            "value": "0"
+          }
+        ]
+      }
+    ]
+  }
+}