--- a/test/rpc-integration/test-collectiontest/state-singleton-map-from-5.execute
+++ b/test/rpc-integration/test-collectiontest/state-singleton-map-from-5.execute
@@ -1,0 +1,1 @@
+../resources/collectiontest/singleton-map.json