… test/rpc-integration/test-collectiontest/response-singleton-map-from-5.json
… test/rpc-integration/test-simplify/response-with-logging.json