fixes #422