@jberthold 2023-12-14
## Kontrol
| Test                                           | 422-handle-in_keys-in-booster-if-possible time | master-6e29375 time | (422-handle-in_keys-in-booster-if-possible/master-6e29375) time 
|------------------------------------------------|------------------------------------------------|---------------------|-----------------------------------------------------------------
| BlockParamsTest.testBlockNumber()              | 21.03                                          | 23.21               | 0.9060749676863421                                              
| AddrTest.test_builtInAddresses()               | 17.99                                          | 19.3                | 0.9321243523316061                                              
| AssertTest.test_assert_true()                  | 49.73                                          | 52.39               | 0.9492269517083413                                              
| ArithmeticTest.test_max1(uint256,uint256)      | 71.89                                          | 69.36               | 1.0364763552479817                                              
| AccountParamsTest.test_GetNonce_true()         | 31.5                                           | 30.35               | 1.0378912685337727                                              
| BlockParamsTest.testChainId(uint256)           | 45.13                                          | 43.39               | 1.0401014058538833                                              
| ExpectRevertTest.testFail_expectRevert_empty() | 32.18                                          | 30.61               | 1.0512904279647175                                              
| LabelTest.testLabel()                          | 34.69                                          | 31.82               | 1.0901948460087993                                              
| TOTAL                                          | 304.14                                         | 300.43              | 1.0123489664813767   

## KEVM
| Test                                     | 422-handle-in_keys-in-booster-if-possible time | master-6e29375 time | (422-handle-in_keys-in-booster-if-possible/master-6e29375) time 
|------------------------------------------|------------------------------------------------|---------------------|-----------------------------------------------------------------
| benchmarks/overflow00-overflow-spec.k    | 52.68                                          | 65.95               | 0.7987869598180439                                              
| bihu/functional-spec.k                   | 116.03                                         | 132.24              | 0.8774198427102238                                              
| erc20/ds/transferFrom-failure-1-c-spec.k | 128.24                                         | 138.66              | 0.9248521563536709                                              
| mcd/cat-file-addr-pass-rough-spec.k      | 76.41                                          | 79.51               | 0.9610111935605583                                              
| mcd/vat-addui-pass-spec.k                | 200.77                                         | 208.67              | 0.9621411798533571                                              
| erc20/hkg/totalSupply-spec.k             | 44.12                                          | 45.76               | 0.9641608391608392                                              
| bihu/forwardToHotWallet-failure-1-spec.k | 71.58                                          | 68.12               | 1.05079271873165                                                
| benchmarks/storagevar00-spec.k           | 51.76                                          | 48.47               | 1.0678770373426862                                              
| bihu/forwardToHotWallet-success-2-spec.k | 171.16                                         | 156.81              | 1.0915120209170333                                              
| erc20/ds/transferFrom-failure-1-d-spec.k | 96.8                                           | 86.36               | 1.1208893006021305                                              
| benchmarks/requires01-a0gt0-spec.k       | 61.98                                          | 50.95               | 1.2164867517173699                                              
| TOTAL                                    | 1071.53                                        | 1081.5              | 0.9907813222376329