@ehildenb 2023-12-14
Did adding K rules not solve this problem? Adding K rules is a more generic solution that will keep the amount of case-specific code in the backened to a minimum. This approach is fine if it gets our goal, but overall it would be nice to not have datastructure specific rules going on.