@jberthold 2023-12-14
> Did adding K rules not solve this problem? Adding K rules is a more generic solution that will keep the amount of case-specific code in the backened to a minimum. This approach is fine if it gets our goal, but overall it would be nice to not have datastructure specific rules going on.

Definitely agree,  this implementation is bespoke and targeted to work around a specific problem. AFAIK we have a better workaround now so we might not use this one.
The integration test that I added here did not succeed with the rules from `MAP-SYMBOLIC`, I could try adding the rules you suggested to see what that does