@jberthold 2023-12-18
The assumed-definedness work-around in kore-rpc works better than this custom built-in, closing for now